requirements.txt
certifi==2020.12.5
chardet==4.0.0
future==0.18.2
idna==2.10
numpy==1.20.1
opencv-contrib-python==4.5.1.48
Pillow==8.1.0
pymongo==3.11.3
requests==2.25.1
six==1.15.0
stomper==0.4.3
urllib3==1.26.3
websocket-client==0.57.0
